♻️refactor: event/todo에서 발행해서 reminder가 소비하기 위해 만든 dto이므로, common.reminder.dto로 중간 dto 위치 이동
