+++ src/main/java/com/project/backend/domain/common/reminder/dto/ReminderDeleted.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.reminder.dto;
+package com.project.backend.domain.common.reminder.dto;
 
 import com.project.backend.domain.reminder.enums.DeletedType;
 import com.project.backend.domain.reminder.enums.TargetType;
