+++ src/main/java/com/project/backend/domain/reminder/handler/ReminderDeletedHandler.java
@@ -1,6 +1,6 @@
 package com.project.backend.domain.reminder.handler;
 
-import com.project.backend.domain.reminder.dto.ReminderDeleted;
+import com.project.backend.domain.common.reminder.dto.ReminderDeleted;
 import com.project.backend.domain.reminder.service.command.ReminderCommandService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
